toiletstats_backend/requirements.txt
Flask==2.2.2
Flask-Cors==4.0.0
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
Werkzeug==2.2.2
